Also move GTFeatures
