+++ src/main/java/com/gregtechceu/gtceu/common/data/worldgen/GTFeatures.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.common.data;
+package com.gregtechceu.gtceu.common.data.worldgen;
 
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.common.worldgen.feature.FluidSproutFeature;
